CAMEL-24125: Refactor OpenAPI spec into ApiDevConsole for cross-runtime support
